Added new properties (linkTitle, weight)
