+++ src/main/notionRequest/getArticles.ts
@@ -2,11 +2,25 @@ import { log, LogTypes } from "../logger";
 import notion from "./client";
 import type { ListBlockChildrenResponse } from "@notionhq/client/build/src/api-endpoints.d";
 import { ListBlockChildrenResponseResults } from "notion-to-md/build/types";
-import { urlize } from "../helpers/string";
-
-interface frontmatterOptions {
-  author: string;
-}
+import {
+  pageTitle,
+  pageAuthor,
+  pageDraft,
+  pagePublishedAt,
+  pageUpdatedAt,
+  pageTags,
+  pageCategory,
+  pageSection,
+  pageToC,
+  pageLegacyAlert,
+  pageSlug,
+  pageUrl,
+  pageDescription,
+  pageLinkTitle,
+  pageWeight,
+  extractExternalUrl,
+  hasPlainText,
+} from "./property";
 
 // Store all metadata in frontMatter
 // Values that are not directly needed to create Hugo pages are stored in the
@@ -27,6 +41,8 @@ export const getPageFrontmatter = async (
     );
   }
   const properties = pageMeta["properties"];
+
+  log(properties, LogTypes.debug);
 
   let frontMatter: any = {
     sys: {
@@ -68,124 +84,18 @@ export const getPageFrontmatter = async (
     frontMatter["images"] = [extractExternalUrl(properties)];
   }
 
+  // Properties expected to be used by Docsy theme, etc
+  //   linkTitle:
+  //   weight:
+  if (pageLinkTitle(properties)) {
+    frontMatter["linkTitle"] = pageLinkTitle(properties);
+  }
+
+  if (pageWeight(properties)) {
+    frontMatter["weight"] = pageWeight(properties);
+  }
+
   return frontMatter as frontMatter;
-};
-
-const pageTitle = (prop: any): string => {
-  return extractPlainText(prop["Name"]);
-};
-const pageAuthor = (prop: any, options: frontmatterOptions): string => {
-  const defaultAuthor = options.author;
-  if (prop["Author"] === undefined) {
-    return defaultAuthor;
-  }
-  const author = extractPlainText(prop["Author"]);
-  if (author) {
-    return author;
-  } else {
-    return defaultAuthor;
-  }
-};
-const pageDraft = (prop: any): boolean => {
-  if (prop["isDraft"] !== undefined) {
-    return isChecked(prop["isDraft"]);
-  }
-  return false;
-};
-const pagePublishedAt = (prop: any): string => {
-  return extractDateTime(prop["PublishedAt"]);
-};
-const pageUpdatedAt = (prop: any): string => {
-  return extractDateTime(prop["UpdatedAt"]);
-};
-const pageTags = (prop: any): string[] => {
-  return extractMultiValues(prop["Tags"]);
-};
-const pageCategory = (prop: any): string[] => {
-  const category = extractPlainText(prop["Category"]);
-  if (category === "") {
-    return [];
-  }
-
-  return [category];
-};
-const pageSection = (prop: any): string => {
-  return extractPlainText(prop["Section"]);
-};
-const pageToC = (prop: any): boolean => {
-  return isChecked(prop["ToC"]);
-};
-const pageLegacyAlert = (prop: any): boolean => {
-  return isChecked(prop["LegacyAlert"]);
-};
-const pageSlug = (prop: any): string => {
-  const slug = extractPlainText(prop["Slug"]);
-  return urlize(slug);
-};
-const pageUrl = (prop: any): string => {
-  const url = extractPlainText(prop["Url"]);
-  return urlize(url);
-};
-const pageDescription = (prop: any): string => {
-  return extractPlainText(prop["Description"]);
-};
-
-const extractPlainText = (prop: any): string => {
-  if (prop["title"]) {
-    return prop["title"][0]["plain_text"];
-  } else if (prop["rich_text"] && prop["rich_text"].length === 1) {
-    return prop["rich_text"][0]["plain_text"];
-  } else if (prop["select"]) {
-    return prop["select"]["name"];
-  }
-  return "";
-};
-
-const extractMultiValues = (prop: any): string[] => {
-  let values: string[] = [];
-  if (prop["multi_select"]) {
-    const tags: any[] = prop["multi_select"];
-    values = tags.map((tag) => tag["name"]);
-    return values;
-  }
-  return [""];
-};
-
-const extractDateTime = (prop: any): string => {
-  if (prop["date"]) {
-    return prop["date"]["start"];
-  }
-  return "";
-};
-
-const extractExternalUrl = (prop: any): string => {
-  if (!prop["Image"]) {
-    return "";
-  }
-  if (prop["Image"]["files"].length === 0) {
-    return "";
-  }
-
-  const file = prop["Image"]["files"][0];
-  if (file["type"] === "external") {
-    return file["external"]["url"];
-  }
-
-  return "";
-};
-
-const isChecked = (prop: any): boolean => {
-  if (prop["type"] === "checkbox") {
-    return prop["checkbox"];
-  }
-  return false;
-};
-
-const hasPlainText = (prop: any): boolean => {
-  if (prop["rich_text"].length > 0) {
-    return true;
-  }
-  return false;
 };
 
 const getArticle = async (pageId: string) => {
